// src/core/models/form.tsx
export type FormType = {
    userName: string;
    lastName: string;
    birthDate: string;
    gender: string;
    email: string;
    isChecked: boolean;
    userNameAccess: string;
    password: string;
    repeatPass: string;
    accountType: string;
    step1: boolean;
    step2: boolean;
    step3: boolean;
    step4: boolean;
    dataError: boolean;
    userLogin: string;
    passLogin: string;
};